# specs/001-mock-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HTTP Mock Management System API
  description: API for managing mock configurations in MystiProxy
  version: 1.0.0
  contact:
    name: MystiProxy Team
servers:
  - url: http://localhost:8080/api/v1
    description: Central Management Server
  - url: http://localhost:9090/api/v1
    description: MystiProxy Local API
...
paths:
  /mocks:
    get:
      tags: [Mocks]
      summary: List mock configurations
      parameters:
        - name: environment
          in: query
          schema:
            type: string
          description: Filter by environment ID
        - name: team
          in: query
          schema:
            type: string
          description: Filter by team ID
        - name: path
          in: query
          schema:
            type: string
          description: Filter by path pattern
        - name: method
          in: query
          schema:
            type: string
          description: Filter by HTTP method
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of mock configurations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockListResponse'
    post:
      tags: [Mocks]
      summary: Create a mock configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockCreateRequest'
      responses:
        '201':
          description: Mock created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockConfiguration'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Mock with same path and method already exists
  /mocks/{id}:
    get:
      tags: [Mocks]
      summary: Get a mock configuration
      parameters:
        - $ref: '#/components/parameters/MockId'
      responses:
        '200':
          description: Mock configuration details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockConfiguration'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Mocks]
      summary: Update a mock configuration
      parameters:
        - $ref: '#/components/parameters/MockId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockUpdateRequest'
      responses:
        '200':
          description: Mock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MockConfiguration'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
    delete:
      tags: [Mocks]
      summary: Delete a mock configuration
      parameters:
        - $ref: '#/components/parameters/MockId'
      responses:
        '204':
          description: Mock deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    MockId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Mock configuration ID
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
    MockConfiguration:
      type: object
      required:
        - id
        - name
        - path
        - method
        - matching_rules
        - response_config
        - source
        - version_vector
        - content_hash
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        path:
          type: string
        method:
          type: string
        team_id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        matching_rules:
          $ref: '#/components/schemas/MatchingRules'
        response_config:
          $ref: '#/components/schemas/ResponseConfig'
        state_config:
          $ref: '#/components/schemas/StateConfig'
        source:
          type: string
          enum: [central, local]
        version_vector:
          type: object
          additionalProperties:
            type: integer
        content_hash:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
    MatchingRules:
      type: object
      properties:
        path_pattern:
          type: string
        path_pattern_type:
          type: string
          enum: [exact, prefix, regex]
        headers:
          type: array
          items:
            $ref: '#/components/schemas/HeaderMatch'
        query_params:
          type: array
          items:
            $ref: '#/components/schemas/QueryParamMatch'
        body:
          $ref: '#/components/schemas/BodyMatch'
    HeaderMatch:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        match_type:
          type: string
          enum: [exact, regex, exists]
          default: exact
    QueryParamMatch:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        match_type:
          type: string
          enum: [exact, regex, exists]
          default: exact
    BodyMatch:
      type: object
      properties:
        json_path:
          type: string
        value:
          type: string
        match_type:
          type: string
          enum: [exact, regex, json_path]
    ResponseConfig:
      type: object
      required: [status]
      properties:
        status:
          type: integer
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          $ref: '#/components/schemas/ResponseBody'
        delay_ms:
          type: integer
    ResponseBody:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [static, template, file, script]
        content:
          type: string
        template_vars:
          type: array
          items:
            $ref: '#/components/schemas/TemplateVar'
    TemplateVar:
      type: object
      required: [name, source]
      properties:
        name:
          type: string
        source:
          type: string
          enum: [path, query, header, body]
        path:
          type: string
    StateConfig:
      type: object
      properties:
        initial_state:
          type: string
        transitions:
          type: array
          items:
            $ref: '#/components/schemas/StateTransition'
    StateTransition:
      type: object
      required: [from_state, to_state]
      properties:
        from_state:
          type: string
        to_state:
          type: string
        trigger:
          $ref: '#/components/schemas/StateTrigger'
        response:
          $ref: '#/components/schemas/ResponseConfig'
    StateTrigger:
      type: object
      properties:
        type:
          type: string
          enum: [request, body, header]
        condition:
          type: string
    MockCreateRequest:
      type: object
      required: [name, path, method, matching_rules, response_config]
      properties:
        name:
          type: string
        path:
          type: string
        method:
          type: string
        team_id:
          type: string
          format: uuid
        environment_id:
          type: string
          format: uuid
        matching_rules:
          $ref: '#/components/schemas/MatchingRules'
        response_config:
          $ref: '#/components/schemas/ResponseConfig'
        state_config:
          $ref: '#/components/schemas/StateConfig'
    MockUpdateRequest:
      type: object
      properties:
        name:
          type: string
        matching_rules:
          $ref: '#/components/schemas/MatchingRules'
        response_config:
          $ref: '#/components/schemas/ResponseConfig'
        state_config:
          $ref: '#/components/schemas/StateConfig'
        version_vector:
          type: object
          additionalProperties:
            type: integer
    MockListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/MockConfiguration'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    ConflictResponse:
      type: object
      properties:
        config_id:
          type: string
          format: uuid
        local_version:
          $ref: '#/components/schemas/MockConfiguration'
        central_version:
          $ref: '#/components/schemas/MockConfiguration'
        detected_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string